Abstract most boilerplate code of service tests to a superclass. Fix a small typo.
